+++ online_voting/voting/views/__init__.py
@@ -1,0 +1,5 @@
+# views/__init__.py
+from .auth import *
+from .events import *
+from .votes import *
+from .profiles import *
